// front/lib/generated/asset_registry/types/assetStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum AssetStatus {
  Draft,
  Active,
  Frozen,
  WoundDown,
}

export type AssetStatusArgs = AssetStatus;

export function getAssetStatusEncoder(): FixedSizeEncoder<AssetStatusArgs> {
  return getEnumEncoder(AssetStatus);
}

export function getAssetStatusDecoder(): FixedSizeDecoder<AssetStatus> {
  return getEnumDecoder(AssetStatus);
}

export function getAssetStatusCodec(): FixedSizeCodec<
  AssetStatusArgs,
  AssetStatus
> {
  return combineCodec(getAssetStatusEncoder(), getAssetStatusDecoder());
}