// front/lib/generated/asset_registry/types/offerStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum OfferStatus {
  Open,
  Filled,
  Cancelled,
  Expired,
}

export type OfferStatusArgs = OfferStatus;

export function getOfferStatusEncoder(): FixedSizeEncoder<OfferStatusArgs> {
  return getEnumEncoder(OfferStatus);
}

export function getOfferStatusDecoder(): FixedSizeDecoder<OfferStatus> {
  return getEnumDecoder(OfferStatus);
}

export function getOfferStatusCodec(): FixedSizeCodec<
  OfferStatusArgs,
  OfferStatus
> {
  return combineCodec(getOfferStatusEncoder(), getOfferStatusDecoder());
}