// front/lib/generated/asset_registry/types/saleStatus.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

export enum SaleStatus {
  Open,
  Closed,
}

export type SaleStatusArgs = SaleStatus;

export function getSaleStatusEncoder(): FixedSizeEncoder<SaleStatusArgs> {
  return getEnumEncoder(SaleStatus);
}

export function getSaleStatusDecoder(): FixedSizeDecoder<SaleStatus> {
  return getEnumDecoder(SaleStatus);
}

export function getSaleStatusCodec(): FixedSizeCodec<
  SaleStatusArgs,
  SaleStatus
> {
  return combineCodec(getSaleStatusEncoder(), getSaleStatusDecoder());
}